change way on control page
