+++ assets/styles/global.scss
@@ -1,0 +1,53 @@
+.control {
+  padding: 30px;
+
+  &-basic {
+    margin: 20px 0px 0px;
+  }
+
+  &-box {
+    margin: 50px 0px 0px;
+    width: 300px;
+  }
+
+  &-input {
+    margin: 8px 0px;
+    width: 150px;
+  }
+
+  &-radio {
+    margin: 10px 0px;
+  }
+
+  &-btn {
+    margin: 20px 0px;
+  }
+
+  &-head {
+    font-size: 24px;
+    margin: 20px 0px 10px;
+    color: black;
+  }
+
+  &-text {
+    margin: 8px 0px 0px;
+    color: black;
+  }
+
+  &-table {
+
+    &-row {
+      display: flex;
+    }
+
+    &-head {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      width: 200px;
+      padding: 10px;
+      border: 0.5px solid black;
+      color: black;
+    }
+  }
+}
