+++ components/control-home.vue
@@ -1,0 +1,30 @@
+<template>
+  <div>
+    <div class="control-head">首頁</div>
+    </div>
+</template>
+
+<script>
+export default {
+  props: {
+
+  },
+  data() {
+    return {
+
+    }
+  },
+  mounted() {
+
+  },
+  methods: {
+
+  }
+}
+</script>
+
+<style lang="scss" scoped>
+
+
+
+</style>
